+++ rdap/__main__.py
@@ -1,7 +1,6 @@
 from __future__ import (absolute_import, division, print_function)
 
 import argparse
-import os
 import sys
 
 import munge