+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 ### Fixed
 ### Changed
+- RdapNotFoundError now inherits from RdapHTTPError instead of LookupError
+
 ### Deprecated
 ### Removed
 ### Security